requirements.txt
streamlit
tensorflow-cpu
Pillow
pandas
numpy